refactor: extract GatesRendering trait and Renderable contract from Component
