+++ src/Core/Contracts/Renderable.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Lattice\Lattice\Core\Contracts;
+
+interface Renderable
+{
+    public function shouldRender(): bool;
+}
